Splitting statistics into separate pages/view models
