+++ src/PayDayWPF/Pages/MarkMeetingPage.xaml
@@ -5,7 +5,7 @@
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
       xmlns:viewModels="clr-namespace:PayDayWPF.ViewModels"
       d:DataContext="{d:DesignInstance viewModels:MarkMeetingViewModel}"
-      mc:Ignorable="d" 
+      mc:Ignorable="d"
       d:DesignHeight="450" d:DesignWidth="800"
       Title="MarkMeetingPage">
