+++ src/PayDayWPF/Pages/StatisticsTab1Page.xaml
@@ -1,0 +1,38 @@
+﻿<Page x:Class="PayDayWPF.Pages.StatisticsTab1Page"
+      xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:viewModels="clr-namespace:PayDayWPF.ViewModels"
+      d:DataContext="{d:DesignInstance viewModels:StatisticsTab1ViewModel}"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="StatisticsTab1Page">
+
+    <Grid Background="#FFE5E5E5">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="65"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="30"/>
+            </Grid.RowDefinitions>
+            <StackPanel Grid.Row="0">
+                <Label Content="{Binding LabelText1}" Padding="5,5,5,1"/>
+                <Label Content="{Binding LabelText1b}" Padding="5,1,5,1"/>
+                <Label Content="{Binding LabelText1c}" Padding="5,1,5,1"/>
+            </StackPanel>
+            <lvc:CartesianChart Grid.Row="1" Series="{Binding SeriesCollection1}" AxisY="{Binding AxesYCollection1}" AxisX="{Binding Labels1}"/>
+            <Grid Grid.Row="2">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="2*"/>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="2*"/>
+                </Grid.ColumnDefinitions>
+                <Button Command="{Binding PreviousCommand}" Grid.Column="0">Previous</Button>
+                <Label Content="{Binding Years}" Grid.Column="1" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+                <Button Command="{Binding NextCommand}" Grid.Column="2">Next</Button>
+            </Grid>
+        </Grid>
+    </Grid>
+</Page>
