+++ src/PayDayWPF/Pages/StatisticsTab1Page.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+using Microsoft.Extensions.DependencyInjection;
+using PayDayWPF.ViewModels;
+
+namespace PayDayWPF.Pages
+{
+    public partial class StatisticsTab1Page : Page
+    {
+        public StatisticsTab1Page()
+        {
+            InitializeComponent();
+            DataContext = MainWindow.ServiceProvider.GetService<StatisticsViewModel>().StatisticsTab1ViewModel;
+        }
+    }
+}
