+++ src/PayDayWPF/Pages/StatisticsTab2Page.xaml
@@ -1,0 +1,23 @@
+﻿<Page x:Class="PayDayWPF.Pages.StatisticsTab2Page"
+      xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:viewModels="clr-namespace:PayDayWPF.ViewModels"
+      d:DataContext="{d:DesignInstance viewModels:StatisticsTab2ViewModel}"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="StatisticsTab2Page">
+
+    <Grid Background="#FFE5E5E5">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="30"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+            <Label Content="{Binding LabelText2}"/>
+            <lvc:CartesianChart Grid.Row="1" Series="{Binding SeriesCollection2}" AxisY="{Binding AxesYCollection2}" AxisX="{Binding Labels2}"/>
+        </Grid>
+    </Grid>
+</Page>
