+++ src/PayDayWPF/Pages/StatisticsTab4Page.xaml
@@ -1,0 +1,32 @@
+﻿<Page x:Class="PayDayWPF.Pages.StatisticsTab4Page"
+      xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:viewModels="clr-namespace:PayDayWPF.ViewModels"
+      d:DataContext="{d:DesignInstance viewModels:StatisticsTab4ViewModel}"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="StatisticsTab4Page">
+
+    <Grid Background="#FFE5E5E5">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="30"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="Auto"/>
+            </Grid.RowDefinitions>
+            <Label Content="{Binding LabelText4}"/>
+            <lvc:CartesianChart Grid.Row="1" Series="{Binding SeriesCollection4}" AxisY="{Binding AxesYCollection4}" AxisX="{Binding Labels4}"/>
+            <Grid Grid.Row="2">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="Auto"/>
+                </Grid.ColumnDefinitions>
+                <Slider Margin="30, 6, 30, 0" Maximum="11" TickPlacement="BottomRight" TickFrequency="1" IsSnapToTickEnabled="True" Value="{Binding MonthsScope}"/>
+                <Label Margin="0, 0, 0, 10" Grid.Column="1" Content="{Binding SliderLabel}"/>
+            </Grid>
+        </Grid>
+    </Grid>
+</Page>
