+++ src/PayDayWPF/Pages/StatisticsTab5Page.xaml
@@ -1,0 +1,23 @@
+﻿<Page x:Class="PayDayWPF.Pages.StatisticsTab5Page"
+      xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:viewModels="clr-namespace:PayDayWPF.ViewModels"
+      d:DataContext="{d:DesignInstance viewModels:StatisticsTab5ViewModel}"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="StatisticsTab5Page">
+
+    <Grid Background="#FFE5E5E5">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="30"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+            <Label Content="{Binding LabelText5}"/>
+            <lvc:CartesianChart Grid.Row="1" Series="{Binding SeriesCollection5}" AxisY="{Binding AxesYCollection5}" AxisX="{Binding Labels5}"/>
+        </Grid>
+    </Grid>
+</Page>
